Made purpose of "docs" folder more clear
The name of the folder docs is really misleading since it's not actually documentation that  is related to this project. It's only a folder that contains all the files  we want embedded in our RAG database, so  rename the docs folder to docs2ingest and analyze  all other documentation (not residing in the docs folder) if they have any  reference to the docs folder and  referencing it with a new name instead.